+++ Tests/BuildMasterAutomationTest/BuildMasterAutomationTest.psm1
@@ -71,7 +71,15 @@ finally
 
 function New-BMTestApplication
 {
-    
+    param(
+        $Session,
+        $CommandPath
+    )
+
+    $Name = Split-Path -Path $CommandPath -Leaf
+    $Name = '{0}.{1}' -f $Name,[IO.Path]::GetRandomFileName()
+
+    return New-BMApplication -Session $Session -Name $Name
 }
 
 function New-BMTestSession