+++ RELEASE_NOTES.md
@@ -3,3 +3,5 @@
  * Created `New-BMSession` function for creating a session to a BuildMaster instance
  * Created `Get-BMApplication` function for getting applications.
  * Created `New-BMApplication` function for creating new applications.
+ * Created `New-BMRelease` function for creating releases.
+ * Created `Get-BMRelease` function for getting releases.